Implement UI feedback for workload completion progression
Closes #74

This implements an interface that shows an annotator encouraging
messages as they complete more workloads. The implementation is
driven by a new `completeCount` field provided on the Workload API
response; on the first step of a new workload (_i.e._, the first
image in a workload AND the first annotation step for that image),
a predetermined message is displayed to the annotator depending on
their progress.

Current messages are for:

- Starting your first workload
- Completing your first workload
- Completing your second workload
- Completing your third workload
- Completing any multiple of 10 workloads

The copy should be considered placeholder, but I did try to avoid
using the actual term "workload," as I felt it connotes that we were
placing a job or a burden on the annotator. Not sure what's best here.

In action:

![ux-feedback-demonstration](https://cloud.githubusercontent.com/assets/442115/24468178/6df3f6ac-1485-11e7-9034-e497381e0019.gif)